+++ src/python/cmuts/internal.py
@@ -29,7 +29,6 @@ from __future__ import annotations
 
 import os
 from dataclasses import dataclass
-from enum import Enum
 from typing import Union
 
 import dask.array as da
@@ -37,16 +36,12 @@ import h5py
 import numpy as np
 from scipy.stats import t as student
 
+from .normalize.schemes import get_norm
+
 # Core datatypes and dataclasses
 
 
 ProbingDatum = Union[np.ndarray, da.Array]
-
-
-class NormScheme(Enum):
-    RAW = 0
-    UBR = 1
-    OUTLIER = 2
 
 
 class DataGroups:
@@ -759,91 +754,6 @@ def _merge_conditions(
     )
 
 
-def _get_norm_scheme(raw: bool, outlier: bool) -> NormScheme:
-    if raw and outlier:
-        raise ValueError("The --raw and --norm-outlier flags are mutually exclusive.")
-
-    if raw:
-        return NormScheme.RAW
-    if outlier:
-        return NormScheme.OUTLIER
-    else:
-        return NormScheme.UBR
-
-
-def _get_norm_raw(
-    data: ProbingData,
-    opts: Opts,
-) -> np.ndarray:
-    """
-    No normalization.
-    """
-
-    return np.ones(1, dtype=data.reactivity.dtype)
-
-
-def _get_norm_percentile(
-    data: ProbingData,
-    opts: Opts,
-) -> np.ndarray:
-    READ_CUTOFF = 100
-    PERCENTILE = 90
-
-    _good_pos = data.mask & (data.reads > READ_CUTOFF)[:, None]
-    high_reactivity = data.reactivity[_good_pos]
-
-    if not high_reactivity.size:
-        return _get_norm_raw(data, opts)
-
-    return np.asarray(np.percentile(high_reactivity.flatten(), PERCENTILE))
-
-
-def _get_norm_outlier(
-    data: ProbingData,
-    opts: Opts,
-) -> np.ndarray:
-    """
-    2-8% Normalization: From top 10%, ignore top 2% and divide by average of
-    remaining 8%.
-    """
-
-    def norm_1d(
-        arr: np.ndarray,
-        low: float = 0.02,
-        high: float = 0.1,
-    ) -> float:
-        arr = arr[~np.isnan(arr)]
-
-        p2 = max(1, round(len(arr) * low) - 1)
-        p10 = max(1, round(len(arr) * high) - 1)
-
-        sarr = np.sort(arr)[::-1]
-        if sarr.size == 0:
-            m: float = 1.0
-        else:
-            m = float(np.mean(sarr[p2 : p10 + 1]))
-
-        return m
-
-    return np.apply_along_axis(norm_1d, axis=1, arr=data.reactivity)
-
-
-def _get_norm(
-    data: ProbingData,
-    opts: Opts,
-) -> np.ndarray:
-    scheme = _get_norm_scheme(opts.raw, opts.outlier)
-
-    if scheme == NormScheme.RAW:
-        return _get_norm_raw(data, opts)
-    if scheme == NormScheme.UBR:
-        return _get_norm_percentile(data, opts)
-    if scheme == NormScheme.OUTLIER:
-        return _get_norm_outlier(data, opts)
-
-    raise ValueError(f"Invalid normalization scheme {scheme}.")
-
-
 def _lengths_from_fasta(file: str) -> da.Array:
     lengths = []
     length = 0
@@ -904,7 +814,7 @@ def compute_reactivity(
 
     # Normalize
 
-    norm = _get_norm(combined, opts)
+    norm = get_norm(combined, opts)
 
     mod.normalize(norm)
     combined.normalize(norm)